refactor: 그룹 관리자 인증 로더 경로 수정
이름이 같은 디렉토리를 제거함
